rename simulator.modules to simulator.models
