+++ tests/context.py
@@ -2,5 +2,5 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
 
-from simulator.modules import Clients, GC, GCI, Request, LoadBalancer, Server, ServerWithGCI
+from simulator.models import Clients, GC, GCI, Request, LoadBalancer, Server, ServerWithGCI
 from config import get_config
